fixed the implicit conversion warning from Softwipe

+++ alignment/alignment.h
@@ -418,7 +418,7 @@ public:
     inline size_t getNSite() {
         // if expected_num_sites is specified -> resizing site_pattern
         if (expected_num_sites > -1)
-            site_pattern.resize(expected_num_sites);
+            site_pattern.resize(static_cast<size_t>(expected_num_sites));
         
         return site_pattern.size();
     }
@@ -431,11 +431,11 @@ public:
     }
 
     inline int getPatternID(int site) {
-        return site_pattern[site];
+        return site_pattern[static_cast<size_t>(site)];
     }
 
     inline Pattern getPattern(int site) {
-        return at(site_pattern[site]);
+        return at(static_cast<size_t>(site_pattern[static_cast<size_t>(site)]));
     }
 
     /**
